Revert "feat: add minimal example core"
This reverts commit b2304b23a0d14118c2e8d609a75757ce46f44992.
